Tests/LibWeb/Text/expected/css/scroll-timeline-auto-duration-progress.txt

animation duration: auto
scale at 25% scroll: 0.25
